+++ src/modules/iframe/components/generator/IframeGenerator.js
